separate build profile
MESA development packages should depends on platform specific headers,
and for the android platform, it's android-frameworks-native-headers
binary package built from this source package. That follows all the
binary packages depending on MESA packages will create an dependency
loop and block bootstrapping. So, as what I've done to
android-system-core, this source package will be divided into two build
profiles:

  - pkg.android-frameworks-native.stage1 (default)
  - pkg.android-frameworks-native.stage2

For stage2, we have so far OpenGL depending packages only but not
limited to.